Moved logger to common folder
