+++ src/m0_facilitator/DatabaseFileHelper.ts
@@ -20,7 +20,7 @@ import path from 'path';
 import sqlite from 'sqlite3';
 
 import Directory from './Directory';
-import Logger from './Logger';
+import Logger from '../common/Logger';
 
 /**
  * It creates db file at the path.
