+++ src/m0_facilitator/bin/facilitator-start.ts
@@ -19,7 +19,7 @@ import commander from 'commander';
 
 import Container from '../Container';
 import Facilitator from '../Facilitator';
-import Logger from '../Logger';
+import Logger from '../../common/Logger';
 import ConfigFactory from '../Config/ConfigFactory';
 
 const facilitatorCmd = commander
