+++ src/m0_facilitator/handlers/redeem_and_unstake/RedeemRequestedHandler.ts
@@ -16,7 +16,7 @@
 
 import BigNumber from 'bignumber.js';
 
-import Logger from '../../Logger';
+import Logger from '../../../common/Logger';
 import MessageTransferRequest from '../../models/MessageTransferRequest';
 import MessageTransferRequestRepository, { RequestType } from '../../repositories/MessageTransferRequestRepository';
 import ContractEntityHandler from '../ContractEntityHandler';
