+++ src/m0_facilitator/handlers/stake_and_mint/StakeIntentConfirmedHandler.ts
@@ -16,7 +16,7 @@
 
 import BigNumber from 'bignumber.js';
 
-import Logger from '../../Logger';
+import Logger from '../../../common/Logger';
 import Message from '../../models/Message';
 import {
   MessageDirection, MessageRepository, MessageStatus, MessageType,
