+++ src/m0_facilitator/handlers/stake_and_mint/StakeRequestedHandler.ts
@@ -16,7 +16,7 @@
 
 import BigNumber from 'bignumber.js';
 
-import Logger from '../../Logger';
+import Logger from '../../../common/Logger';
 import MessageTransferRequest from '../../models/MessageTransferRequest';
 import MessageTransferRequestRepository, { RequestType } from '../../repositories/MessageTransferRequestRepository';
 import ContractEntityHandler from '../ContractEntityHandler';
