+++ src/m0_facilitator/services/redeem_and_unstake/ConfirmRedeemIntentService.ts
@@ -25,7 +25,7 @@ import { ProofGenerator } from '@openst/mosaic-proof';
 
 import { EIP20Gateway } from '@openst/mosaic-contracts/dist/interacts/EIP20Gateway';
 import { ORIGIN_GAS_PRICE } from '../../Constants';
-import Logger from '../../Logger';
+import Logger from '../../../common/Logger';
 import Gateway from '../../models/Gateway';
 import Message from '../../models/Message';
 import MessageTransferRequest from '../../models/MessageTransferRequest';
