+++ src/m0_facilitator/services/stake_and_mint/ConfirmStakeIntentService.ts
@@ -26,7 +26,7 @@ import { EIP20CoGateway } from '@openst/mosaic-contracts/dist/interacts/EIP20CoG
 import { ProofGenerator } from '@openst/mosaic-proof';
 
 import { AUXILIARY_GAS_PRICE } from '../../Constants';
-import Logger from '../../Logger';
+import Logger from '../../../common/Logger';
 import Gateway from '../../models/Gateway';
 import Message from '../../models/Message';
 import MessageTransferRequest from '../../models/MessageTransferRequest';
