+++ src/m0_facilitator/subscriptions/GraphClient.ts
@@ -26,7 +26,7 @@ import fetch from 'node-fetch';
 import { SubscriptionClient } from 'subscriptions-transport-ws';
 import WebSocket from 'ws';
 
-import Logger from '../Logger';
+import Logger from '../../common/Logger';
 import ContractEntity, { EntityType } from '../models/ContractEntity';
 import ContractEntityRepository from '../repositories/ContractEntityRepository';
 import TransactionHandler from '../TransactionHandler';
