+++ test_integration/m0_facilitator/helpers/EIP20TokenHelper.ts
@@ -21,7 +21,7 @@ import { UtilityToken } from '@openst/mosaic-contracts/dist/interacts/UtilityTok
 
 import { HelperInterface } from './HelperInterface';
 import Utils from '../Utils';
-import Logger from '../../../src/m0_facilitator/Logger';
+import Logger from '../../../src/common/Logger';
 
 export default class EIP20TokenHelper implements HelperInterface {
